backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator>=2.0.0
Pillow
pyzbar
